+++ resources/store.py
@@ -1,17 +1,17 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 
 from models.store import StoreModel
 
 
 class StoreList(Resource):
-    @jwt_required()
+    @jwt_required
     def get(self):
         return { 'stores': [ store.json() for store in StoreModel.find_all() ] }
 
 
 class Store(Resource):
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         store = StoreModel.find_by_name(name)
 
@@ -19,7 +19,7 @@ class Store(Resource):
             return store.json()
         return { 'message': None }, 404
 
-    @jwt_required()
+    @jwt_required
     def post(self, name):
         if StoreModel.find_by_name(name):
             return { 'message': f'Ya existe {name}' }, 400
@@ -32,11 +32,15 @@ class Store(Resource):
 
         return store.json(), 201
 
-    @jwt_required()
+    @jwt_required
     def delete(self, name):
-        store = StoreModel.find_by_name(name)
-        if store:
-            store.delete_store()
-            return { 'message': f'{name} fue borrado exitosamente' }, 200
+        claims = get_jwt_claims()
+        if claims['is_admin']:
+            store = StoreModel.find_by_name(name)
+            if store:
+                store.delete_store()
+                return { 'message': f'{name} fue borrado exitosamente' }, 200
 
-        return { 'message': f'{name} no se encontró o ya fue borrado' }, 400
+            return { 'message': f'{name} no se encontró o ya fue borrado' }, 400
+
+        return { 'message': 'Debes ser admin' }, 401
